requirements.txt
pefile
requests
yara-python
Pillow
fpdf2
ssdeep
flask
flask-cors
customtkinter
